+++ .vitepress/theme/components/CommentUtteranc.vue
@@ -5,6 +5,7 @@ const utterancesRef = ref()
 const { theme, isDark } = useData()
 onMounted(() => {
     nextTick(() => {
+        //通过theme.value.comment的配置取得
         let { repo, issueTerm = 'pathname' } = theme.value.comment
         if (repo) {
             let utterances = document.createElement('script')